pyproject.toml
[project]
name = "poc-dashboard"
version = "0.1.0"
description = "poc dashboard connecting to k3s data for local dev"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=3.1.1",
    "duckdb>=1.3.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
